requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
llama-cpp-python==0.3.16
pydantic==2.12.5